+++ lightly/api/upload.py
@@ -129,103 +129,6 @@ def _upload_single_image(image,
     return success
 
 
-def upload_dataset(dataset: LightlyDataset,
-                   dataset_id: str,
-                   token: str,
-                   max_workers: int = 8,
-                   mode: str = 'thumbnails',
-                   verbose: bool = True):
-    """Uploads images from a directory to the Lightly cloud solution.
-
-    Args:
-        dataset
-            The dataset to upload.
-        dataset_id:
-            The unique identifier for the dataset.
-        token:
-            Token for authentication.
-        max_workers:
-            Maximum number of workers uploading images in parallel.
-        max_requests:
-            Maximum number of requests a single worker can do before he has
-            to wait for the others.
-        mode:
-            One of [full, thumbnails, metadata]. Whether to upload thumbnails,
-            full images, or metadata only.
-
-    Raises:
-        ValueError if dataset is too large.
-        RuntimeError if the connection to the server failed.
-        RuntimeError if dataset already has an initial tag.
-
-    """
-
-    # check the allowed dataset size
-    api_max_dataset_size, status_code = routes.users.get_quota(token)
-    max_dataset_size = min(api_max_dataset_size, LIGHTLY_MAXIMUM_DATASET_SIZE)
-    if len(dataset) > max_dataset_size:
-        msg = f'Your dataset has {len(dataset)} samples which'
-        msg += f' is more than the allowed maximum of {max_dataset_size}'
-        raise ValueError(msg)
-
-    # check whether connection to server was possible
-    if status_code != 200:
-        msg = f'Connection to server failed with status code {status_code}.'
-        raise RuntimeError(msg)
-
-    # handle the case where len(dataset) < max_workers
-    max_workers = min(len(dataset), max_workers)
-
-    # upload the samples
-    if verbose:
-        print(f'Uploading images (with {max_workers} workers).', flush=True)
-
-    pbar = tqdm.tqdm(unit='imgs', total=len(dataset))
-    tqdm_lock = tqdm.tqdm.get_lock()
-
-    # define lambda function for concurrent upload
-    def lambda_(i):
-        # load image
-        image, label, filename = dataset[i]
-        # upload image
-        success = _upload_single_image(
-            image=image,
-            label=label,
-            filename=filename,
-            dataset_id=dataset_id,
-            token=token,
-            mode=mode,
-        )
-        # update the progress bar
-        tqdm_lock.acquire()  # lock
-        pbar.update(1)  # update
-        tqdm_lock.release()  # unlock
-        # return whether the upload was successful
-        return success
-
-    with ThreadPoolExecutor(max_workers=max_workers) as executor:
-        results = list(executor.map(
-            lambda_, [i for i in range(len(dataset))], chunksize=1))
-
-    if not all(results):
-        msg = 'Warning: Unsuccessful upload(s)! '
-        msg += 'This could cause problems when uploading embeddings.'
-        msg += 'Failed at image: {}'.format(results.index(False))
-        warnings.warn(msg)
-
-    # set image type of data and create initial tag
-    if mode == 'full':
-        img_type = 'full'
-    elif mode == 'thumbnails':
-        img_type = 'thumbnail'
-    else:
-        img_type = "meta"
-
-    api_client = get_api_client(token=token)
-    tags_api = TagsApi(api_client=api_client)
-    initial_tag_create_request = InitialTagCreateRequest(img_type=img_type)
-    tags_api.create_initial_tag_by_dataset_id(body=initial_tag_create_request, dataset_id=dataset_id)
-
 
 def upload_file_with_signed_url(file, url: str) -> bool:
     """Upload a file to the cloud storage using a signed URL.
